requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Document processing (alternatives to avoid compilation issues)
pdfplumber==0.11.0
pypdf2==3.0.1
python-docx==1.1.0
pdfminer.six==20231228

# Data processing
pandas==2.1.3

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests (for AI API calls)
requests==2.31.0
httpx==0.25.2

# Background tasks (optional for now)
celery==5.3.4
redis==5.0.1

# Cloud storage (for production)
boto3==1.34.0